[Major] Implemented groups in privilege. Refactored nearly everything

+++ privilege/src/test/java/li/strolch/privilege/test/XmlTest.java
@@ -16,7 +16,7 @@
 package li.strolch.privilege.test;
 
 import li.strolch.privilege.handler.*;
-import li.strolch.privilege.model.IPrivilege;
+import li.strolch.privilege.model.Privilege;
 import li.strolch.privilege.model.UserState;
 import li.strolch.privilege.model.internal.*;
 import li.strolch.privilege.test.model.DummySsoHandler;
@@ -76,6 +76,11 @@ public class XmlTest {
 		}
 
 		tmpFile = new File(TARGET_TEST + "PrivilegeUsersTest.xml");
+		if (tmpFile.exists() && !tmpFile.delete()) {
+			throw new RuntimeException("Tmp still exists and can not be deleted at " + tmpFile.getAbsolutePath());
+		}
+
+		tmpFile = new File(TARGET_TEST + "PrivilegeGroupsTest.xml");
 		if (tmpFile.exists() && !tmpFile.delete()) {
 			throw new RuntimeException("Tmp still exists and can not be deleted at " + tmpFile.getAbsolutePath());
 		}
@@ -194,12 +199,13 @@ public class XmlTest {
 		assertEquals("1", admin.getUserId());
 		assertEquals("admin", admin.getUsername());
 		assertEquals("cb69962946617da006a2f95776d78b49e5ec7941d2bdb2d25cdb05f957f64344",
-				StringHelper.toHexString(admin.getPasswordCrypt().getPassword()));
-		assertEquals("61646d696e", StringHelper.toHexString(admin.getPasswordCrypt().getSalt()));
+				StringHelper.toHexString(admin.getPasswordCrypt().password()));
+		assertEquals("61646d696e", StringHelper.toHexString(admin.getPasswordCrypt().salt()));
 		assertEquals("Application", admin.getFirstname());
 		assertEquals("Administrator", admin.getLastname());
 		assertEquals(UserState.ENABLED, admin.getUserState());
 		assertEquals("en-GB", admin.getLocale().toLanguageTag());
+		assertEquals(Set.of("GroupA"), admin.getGroups());
 		MatcherAssert.assertThat(admin.getRoles(), containsInAnyOrder("PrivilegeAdmin", "AppUser"));
 		Map<String, String> properties = admin.getProperties();
 		assertEquals(new HashSet<>(Arrays.asList("organization", "organizationalUnit")), properties.keySet());
@@ -215,8 +221,48 @@ public class XmlTest {
 		assertEquals("Administrator", systemAdmin.getLastname());
 		assertEquals(UserState.SYSTEM, systemAdmin.getUserState());
 		assertEquals("en-GB", systemAdmin.getLocale().toLanguageTag());
+		assertEquals(Set.of(), systemAdmin.getGroups());
 		MatcherAssert.assertThat(systemAdmin.getRoles(), containsInAnyOrder("system_admin_privileges"));
 		assertTrue(systemAdmin.getProperties().isEmpty());
+
+		// admin2
+		User admin2 = findUser("admin2", users);
+		assertEquals("1", admin2.getUserId());
+		assertEquals("admin2", admin2.getUsername());
+		assertEquals("8c6976e5b5410415bde908bd4dee15dfb167a9c873fc4bb8a81f6f2ab448a918",
+				StringHelper.toHexString(admin2.getPasswordCrypt().password()));
+		assertEquals("Application", admin2.getFirstname());
+		assertEquals("Administrator", admin2.getLastname());
+		assertEquals(UserState.ENABLED, admin2.getUserState());
+		assertEquals("en-GB", admin2.getLocale().toLanguageTag());
+		MatcherAssert.assertThat(admin2.getGroups(), containsInAnyOrder("AppUserLocationA"));
+		MatcherAssert.assertThat(admin2.getRoles(), containsInAnyOrder("PrivilegeAdmin"));
+		properties = admin2.getProperties();
+		assertEquals(new HashSet<>(Arrays.asList("organization", "organizationalUnit")), properties.keySet());
+		assertEquals("eitchnet.ch", properties.get("organization"));
+		assertEquals("Development", properties.get("organizationalUnit"));
+
+	}
+
+	@Test
+	public void canReadGroups() {
+
+		PrivilegeGroupsSaxReader xmlHandler = new PrivilegeGroupsSaxReader();
+		File xmlFile = new File(SRC_TEST + "PrivilegeGroups.xml");
+		XmlHelper.parseDocument(xmlFile, xmlHandler);
+
+		Map<String, Group> groups = xmlHandler.getGroups();
+		assertNotNull(groups);
+
+		assertEquals(1, groups.size());
+
+		// group AppUserLocationA
+		Group group = groups.get("AppUserLocationA");
+		assertEquals("AppUserLocationA", group.name());
+		MatcherAssert.assertThat(group.roles(), containsInAnyOrder("AppUser", "ModelAccessor", "UserPrivileges"));
+		Map<String, String> properties = group.getProperties();
+		assertEquals(new HashSet<>(List.of("location")), properties.keySet());
+		assertEquals("LocationA", properties.get("location"));
 	}
 
 	@Test
@@ -241,18 +287,18 @@ public class XmlTest {
 		Role privilegeAdmin = findRole("PrivilegeAdmin", roles);
 		assertEquals("PrivilegeAdmin", privilegeAdmin.getName());
 		assertEquals(18, privilegeAdmin.getPrivilegeNames().size());
-		IPrivilege privilegeAction = privilegeAdmin.getPrivilege(PrivilegeHandler.PRIVILEGE_ACTION);
+		Privilege privilegeAction = privilegeAdmin.getPrivilege(PrivilegeHandler.PRIVILEGE_ACTION);
 		assertFalse(privilegeAction.isAllAllowed());
 		assertEquals(5, privilegeAction.getAllowList().size());
 		assertEquals(0, privilegeAction.getDenyList().size());
 		assertEquals("DefaultPrivilege", privilegeAction.getPolicy());
 
-		IPrivilege privilegeAddRole = privilegeAdmin.getPrivilege(PrivilegeHandler.PRIVILEGE_ADD_ROLE);
+		Privilege privilegeAddRole = privilegeAdmin.getPrivilege(PrivilegeHandler.PRIVILEGE_ADD_ROLE);
 		assertTrue(privilegeAddRole.isAllAllowed());
 		assertEquals(0, privilegeAddRole.getAllowList().size());
 		assertEquals(0, privilegeAddRole.getDenyList().size());
 
-		IPrivilege privilegeRemRoleFromUser = privilegeAdmin.getPrivilege(
+		Privilege privilegeRemRoleFromUser = privilegeAdmin.getPrivilege(
 				PrivilegeHandler.PRIVILEGE_REMOVE_ROLE_FROM_USER);
 		assertTrue(privilegeRemRoleFromUser.isAllAllowed());
 		assertEquals(0, privilegeRemRoleFromUser.getAllowList().size());
@@ -264,7 +310,7 @@ public class XmlTest {
 		assertEquals(new HashSet<>(Collections.singletonList("li.strolch.privilege.test.model.TestRestrictable")),
 				appUser.getPrivilegeNames());
 
-		IPrivilege testRestrictable = appUser.getPrivilege("li.strolch.privilege.test.model.TestRestrictable");
+		Privilege testRestrictable = appUser.getPrivilege("li.strolch.privilege.test.model.TestRestrictable");
 		assertEquals("li.strolch.privilege.test.model.TestRestrictable", testRestrictable.getName());
 		assertEquals("DefaultPrivilege", testRestrictable.getPolicy());
 		assertTrue(testRestrictable.isAllAllowed());
@@ -279,7 +325,7 @@ public class XmlTest {
 				containsInAnyOrder("li.strolch.privilege.handler.SystemAction",
 						"li.strolch.privilege.test.model.TestSystemRestrictable"));
 
-		IPrivilege testSystemUserAction = systemAdminPrivileges.getPrivilege(
+		Privilege testSystemUserAction = systemAdminPrivileges.getPrivilege(
 				"li.strolch.privilege.handler.SystemAction");
 		assertEquals("li.strolch.privilege.handler.SystemAction", testSystemUserAction.getName());
 		assertEquals("DefaultPrivilege", testSystemUserAction.getPolicy());
@@ -287,7 +333,7 @@ public class XmlTest {
 		assertEquals(1, testSystemUserAction.getAllowList().size());
 		assertEquals(1, testSystemUserAction.getDenyList().size());
 
-		IPrivilege testSystemRestrictable = systemAdminPrivileges.getPrivilege(
+		Privilege testSystemRestrictable = systemAdminPrivileges.getPrivilege(
 				"li.strolch.privilege.test.model.TestSystemRestrictable");
 		assertEquals("li.strolch.privilege.test.model.TestSystemRestrictable", testSystemRestrictable.getName());
 		assertEquals("DefaultPrivilege", testSystemRestrictable.getPolicy());
@@ -302,7 +348,7 @@ public class XmlTest {
 		MatcherAssert.assertThat(restrictedRole.getPrivilegeNames(),
 				containsInAnyOrder("li.strolch.privilege.handler.SystemAction"));
 
-		IPrivilege testSystemUserAction2 = restrictedRole.getPrivilege("li.strolch.privilege.handler.SystemAction");
+		Privilege testSystemUserAction2 = restrictedRole.getPrivilege("li.strolch.privilege.handler.SystemAction");
 		assertEquals("li.strolch.privilege.handler.SystemAction", testSystemUserAction2.getName());
 		assertEquals("DefaultPrivilege", testSystemUserAction2.getPolicy());
 		assertFalse(testSystemUserAction2.isAllAllowed());
@@ -326,29 +372,35 @@ public class XmlTest {
 	public void canWriteUsers() throws XMLStreamException, IOException {
 
 		Map<String, String> propertyMap;
+		Set<String> groups;
 		Set<String> userRoles;
 
 		List<User> users = new ArrayList<>();
 		propertyMap = new HashMap<>();
 		propertyMap.put("prop1", "value1");
+		groups = new HashSet<>();
+		groups.add("group1");
 		userRoles = new HashSet<>();
 		userRoles.add("role1");
-		UserHistory history = new UserHistory();
-		history.setFirstLogin(ZonedDateTime.of(LocalDateTime.of(2020, 1, 2, 2, 3, 4, 5), ZoneId.systemDefault()));
+		UserHistory history = UserHistory.EMPTY.withFirstLogin(
+				ZonedDateTime.of(LocalDateTime.of(2020, 1, 2, 2, 3, 4, 5), ZoneId.systemDefault()));
 		User user1 = new User("1", "user1",
 				new PasswordCrypt("blabla".getBytes(), "blabla".getBytes(), "PBKDF2WithHmacSHA512", 10000, 256), "Bob",
-				"White", UserState.DISABLED, userRoles, Locale.ENGLISH, propertyMap, false, history);
+				"White", UserState.DISABLED, groups, userRoles, Locale.ENGLISH, propertyMap, false, history);
 		users.add(user1);
 
 		propertyMap = new HashMap<>();
 		propertyMap.put("prop2", "value2");
+		groups = new HashSet<>();
+		groups.add("group2");
 		userRoles = new HashSet<>();
 		userRoles.add("role2");
-		history = new UserHistory();
-		history.setFirstLogin(ZonedDateTime.of(LocalDateTime.of(2020, 1, 2, 2, 3, 4, 5), ZoneId.systemDefault()));
-		history.setLastLogin(ZonedDateTime.of(LocalDateTime.of(2020, 1, 5, 2, 3, 4, 5), ZoneId.systemDefault()));
+		history = UserHistory.EMPTY.withFirstLogin(
+						ZonedDateTime.of(LocalDateTime.of(2020, 1, 2, 2, 3, 4, 5), ZoneId.systemDefault()))
+				.withLastLogin(ZonedDateTime.of(LocalDateTime.of(2020, 1, 5, 2, 3, 4, 5), ZoneId.systemDefault()));
 		User user2 = new User("2", "user2", new PasswordCrypt("haha".getBytes(), "haha".getBytes(), null, -1, -1),
-				"Leonard", "Sheldon", UserState.ENABLED, userRoles, Locale.ENGLISH, propertyMap, false, history);
+				"Leonard", "Sheldon", UserState.ENABLED, groups, userRoles, Locale.ENGLISH, propertyMap, false,
+				history);
 		users.add(user2);
 
 		File modelFile = new File(TARGET_TEST + "PrivilegeUsersTest.xml");
@@ -370,8 +422,8 @@ public class XmlTest {
 		assertEquals(user1.getFirstname(), parsedUser1.getFirstname());
 		assertEquals(user1.getLastname(), parsedUser1.getLastname());
 		assertEquals(user1.getLocale(), parsedUser1.getLocale());
-		assertArrayEquals(user1.getPasswordCrypt().getPassword(), parsedUser1.getPasswordCrypt().getPassword());
-		assertArrayEquals(user1.getPasswordCrypt().getSalt(), parsedUser1.getPasswordCrypt().getSalt());
+		assertArrayEquals(user1.getPasswordCrypt().password(), parsedUser1.getPasswordCrypt().password());
+		assertArrayEquals(user1.getPasswordCrypt().salt(), parsedUser1.getPasswordCrypt().salt());
 		assertEquals(user1.getProperties(), parsedUser1.getProperties());
 		assertEquals(user1.getUserId(), parsedUser1.getUserId());
 		assertEquals(user1.getUserState(), parsedUser1.getUserState());
@@ -380,8 +432,8 @@ public class XmlTest {
 		assertEquals(user2.getFirstname(), parsedUser2.getFirstname());
 		assertEquals(user2.getLastname(), parsedUser2.getLastname());
 		assertEquals(user2.getLocale(), parsedUser2.getLocale());
-		assertArrayEquals(user2.getPasswordCrypt().getPassword(), parsedUser2.getPasswordCrypt().getPassword());
-		assertArrayEquals(user2.getPasswordCrypt().getSalt(), parsedUser2.getPasswordCrypt().getSalt());
+		assertArrayEquals(user2.getPasswordCrypt().password(), parsedUser2.getPasswordCrypt().password());
+		assertArrayEquals(user2.getPasswordCrypt().salt(), parsedUser2.getPasswordCrypt().salt());
 		assertEquals(user2.getProperties(), parsedUser2.getProperties());
 		assertEquals(user2.getUserId(), parsedUser2.getUserId());
 		assertEquals(user2.getUserState(), parsedUser2.getUserState());
@@ -389,13 +441,48 @@ public class XmlTest {
 	}
 
 	@Test
+	public void canWriteGroups() throws XMLStreamException, IOException {
+
+		Map<String, String> propertyMap;
+		Set<String> roles;
+
+		List<Group> groups = new ArrayList<>();
+		propertyMap = new HashMap<>();
+		propertyMap.put("prop1", "value1");
+		roles = new HashSet<>();
+		roles.add("role1");
+		Group newGroup = new Group("group1", roles, propertyMap);
+		groups.add(newGroup);
+
+		File modelFile = new File(TARGET_TEST + "PrivilegeGroupsTest.xml");
+		PrivilegeGroupsSaxWriter configSaxWriter = new PrivilegeGroupsSaxWriter(groups, modelFile);
+		configSaxWriter.write();
+
+		PrivilegeGroupsSaxReader xmlHandler = new PrivilegeGroupsSaxReader();
+		XmlHelper.parseDocument(modelFile, xmlHandler);
+
+		Map<String, Group> parsedGroups = xmlHandler.getGroups();
+		assertNotNull(parsedGroups);
+		assertEquals(1, parsedGroups.size());
+
+		// group group1
+		Group parsedGroup1 = parsedGroups.get("group1");
+		assertNotNull(parsedGroup1);
+		assertEquals("group1", parsedGroup1.name());
+		MatcherAssert.assertThat(parsedGroup1.roles(), containsInAnyOrder("role1"));
+		Map<String, String> properties = parsedGroup1.getProperties();
+		assertEquals(new HashSet<>(List.of("prop1")), properties.keySet());
+		assertEquals("value1", properties.get("prop1"));
+	}
+
+	@Test
 	public void canWriteRoles() throws XMLStreamException, IOException {
 
-		Map<String, IPrivilege> privilegeMap;
+		Map<String, Privilege> privilegeMap;
 		List<Role> roles = new ArrayList<>();
 		Set<String> list = Collections.emptySet();
 		privilegeMap = new HashMap<>();
-		privilegeMap.put("priv1", new PrivilegeImpl("priv1", "DefaultPrivilege", true, list, list));
+		privilegeMap.put("priv1", new Privilege("priv1", "DefaultPrivilege", true, list, list));
 		Role role1 = new Role("role1", privilegeMap);
 		roles.add(role1);
 
@@ -404,7 +491,7 @@ public class XmlTest {
 		denyList.add("myself");
 		Set<String> allowList = new HashSet<>();
 		allowList.add("other");
-		privilegeMap.put("priv2", new PrivilegeImpl("priv2", "DefaultPrivilege", false, denyList, allowList));
+		privilegeMap.put("priv2", new Privilege("priv2", "DefaultPrivilege", false, denyList, allowList));
 		Role role2 = new Role("role2", privilegeMap);
 		roles.add(role2);
 
@@ -428,8 +515,8 @@ public class XmlTest {
 		Set<String> privilegeNames = role1.getPrivilegeNames();
 		assertEquals(privilegeNames, parsedRole1.getPrivilegeNames());
 		for (String privilegeName : privilegeNames) {
-			IPrivilege privilege = role1.getPrivilege(privilegeName);
-			IPrivilege privilege2 = parsedRole1.getPrivilege(privilegeName);
+			Privilege privilege = role1.getPrivilege(privilegeName);
+			Privilege privilege2 = parsedRole1.getPrivilege(privilegeName);
 			assertNotNull(privilege);
 			assertNotNull(privilege2);
 
@@ -443,8 +530,8 @@ public class XmlTest {
 		assertEquals(role2.getPrivilegeNames(), parsedRole2.getPrivilegeNames());
 		privilegeNames = role2.getPrivilegeNames();
 		for (String privilegeName : privilegeNames) {
-			IPrivilege privilege = role2.getPrivilege(privilegeName);
-			IPrivilege privilege2 = parsedRole2.getPrivilege(privilegeName);
+			Privilege privilege = role2.getPrivilege(privilegeName);
+			Privilege privilege2 = parsedRole2.getPrivilege(privilegeName);
 			assertNotNull(privilege);
 			assertNotNull(privilege2);
